requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asgiref=3.5.2=py310h06a4308_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.11.17=hbcca054_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py310h06a4308_0
django=4.1=py310h06a4308_0
djangorestframework=3.14.0=pyhd8ed1ab_0
idna=3.4=pypi_0
krb5=1.20.1=h143b758_1
ld_impl_linux-64=2.38=h1181459_1
libedit=3.1.20221030=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_3
libgomp=13.2.0=h807b86a_3
libpq=12.15=hdbd6064_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=3.1.4=hd590300_0
pip=23.3=py310h06a4308_0
psycopg2=2.9.3=py310h5eee18b_1
python=3.10.13=h955ad1f_0
python-dotenv=0.21.0=py310h06a4308_0
pytz=2023.3.post1=pyhd8ed1ab_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
setuptools=68.0.0=py310h06a4308_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=py310h06a4308_0
stripe=7.6.0=pypi_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.8.0=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=2.1.0=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
